Use a circular buffer for input from the device driver
This speeds up things a little bit, and makes the behavior in the case
of an overflow a little bit better. In addition, because we can rely on
one thread updating the head, and another updating the tail, if we only
access each one from it's respective consumer's thread we can read and
write to/from the buffer concurrently without having to use any locks.